Enormously simplified the client code by reusing a lot of the service classes. Removed the JSON codec because it was adding noise, and now use pretty pretty Stream stuff to create and parse a list of moods. Had to change happiness chart to a double because despite my best efforts, that's how the data comes out.

+++ dashboard/src/main/java/com/mechanitis/demo/sense/client/mood/HappinessController.java
@@ -11,7 +11,7 @@ import java.util.ResourceBundle;
 
 public class HappinessController implements Initializable {
     @FXML private NumberAxis yAxis;
-    @FXML private BarChart<String, Long> happinessOverTime;
+    @FXML private BarChart<String, Double> happinessOverTime;
 
     public void setData(HappinessChartData data) {
         happinessOverTime.getData().add(data.getDataSeries());
